+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 
+* New argument `--remote` to specify remote (default: 'origin')
+
 ### Changed
 
 ### Deprecated